imageSplitCatalog/requirements.txt
boto3
urllib3
pillow
picamraw
tifffile